requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-cloud-vision
google-cloud-aiplatform
httpx
